requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
urllib3==1.26.12
pandas==1.5.1
PyYAML==6.0
nltk==3.7
scikit-learn==1.1.3
gitpython==3.1.29
